refactor: change order-details state file names to order-confirmation
